requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.1.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
